Move InvocationMatcher out create compound matcher
