+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/bestpractices/JUnitAssertionsShouldIncludeMessageRule.java
@@ -4,31 +4,28 @@
 
 package net.sourceforge.pmd.lang.java.rule.bestpractices;
 
-import static net.sourceforge.pmd.util.CollectionUtil.listOf;
-
-import java.util.List;
-
 import net.sourceforge.pmd.lang.java.ast.ASTMethodCall;
 import net.sourceforge.pmd.lang.java.rule.AbstractJavaRulechainRule;
 import net.sourceforge.pmd.lang.java.rule.internal.TestFrameworksUtil;
-import net.sourceforge.pmd.lang.java.types.TypeTestUtil.InvocationMatcher;
+import net.sourceforge.pmd.lang.java.types.InvocationMatcher;
+import net.sourceforge.pmd.lang.java.types.InvocationMatcher.CompoundInvocationMatcher;
 
 public class JUnitAssertionsShouldIncludeMessageRule extends AbstractJavaRulechainRule {
 
-    private final List<InvocationMatcher> checks =
-        listOf(
-            InvocationMatcher.parse("_#assertEquals(_,_)"),
-            InvocationMatcher.parse("_#assertTrue(_)"),
-            InvocationMatcher.parse("_#assertFalse(_)"),
-            InvocationMatcher.parse("_#assertSame(_,_)"),
-            InvocationMatcher.parse("_#assertNotSame(_,_)"),
-            InvocationMatcher.parse("_#assertNull(_)"),
-            InvocationMatcher.parse("_#assertNotNull(_)"),
-            InvocationMatcher.parse("_#assertArrayEquals(_,_)"),
-            InvocationMatcher.parse("_#assertThat(_,_)"),
-            InvocationMatcher.parse("_#fail()"),
-            InvocationMatcher.parse("_#assertEquals(float,float,float)"),
-            InvocationMatcher.parse("_#assertEquals(double,double,double)")
+    private final CompoundInvocationMatcher checks =
+        InvocationMatcher.parseAll(
+            "_#assertEquals(_,_)",
+            "_#assertTrue(_)",
+            "_#assertFalse(_)",
+            "_#assertSame(_,_)",
+            "_#assertNotSame(_,_)",
+            "_#assertNull(_)",
+            "_#assertNotNull(_)",
+            "_#assertArrayEquals(_,_)",
+            "_#assertThat(_,_)",
+            "_#fail()",
+            "_#assertEquals(float,float,float)",
+            "_#assertEquals(double,double,double)"
         );
 
     public JUnitAssertionsShouldIncludeMessageRule() {
@@ -38,11 +35,8 @@ public class JUnitAssertionsShouldIncludeMessageRule extends AbstractJavaRulecha
     @Override
     public Object visit(ASTMethodCall node, Object data) {
         if (TestFrameworksUtil.isCallOnAssertionContainer(node)) {
-            for (InvocationMatcher check : checks) {
-                if (check.matchesCall(node)) {
-                    addViolation(data, node);
-                    break;
-                }
+            if (checks.anyMatch(node)) {
+                addViolation(data, node);
             }
         }
         return null;
